# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APSO Backend API",
    "description": "AI-powered SDLC intelligence backend. MVP scope is read-only extraction and analysis.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/engines/definition-gap/analyze": {
      "post": {
        "tags": [
          "engines"
        ],
        "summary": "Analyze Definition Gap",
        "operationId": "analyze_definition_gap_api_v1_engines_definition_gap_analyze_post",
        "parameters": [
          {
            "name": "use_llm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Use Llm"
            }
          },
          {
            "name": "persist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Persist"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementGapInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementGapOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/engines/code-quality/analyze": {
      "post": {
        "tags": [
          "engines"
        ],
        "summary": "Analyze Code Quality",
        "operationId": "analyze_code_quality_api_v1_engines_code_quality_analyze_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeQualityAnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeQualityAnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/engines/coverage/verify": {
      "post": {
        "tags": [
          "engines"
        ],
        "summary": "Verify Coverage",
        "operationId": "verify_coverage_api_v1_engines_coverage_verify_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverageVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/tenants": {
      "get": {
        "tags": [
          "core"
        ],
        "summary": "List Tenants",
        "operationId": "list_tenants_api_v1_core_tenants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantRead"
                  },
                  "type": "array",
                  "title": "Response List Tenants Api V1 Core Tenants Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "core"
        ],
        "summary": "Create Tenant",
        "operationId": "create_tenant_api_v1_core_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/projects": {
      "post": {
        "tags": [
          "core"
        ],
        "summary": "Create Project",
        "operationId": "create_project_api_v1_core_projects_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "core"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_api_v1_core_projects_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectRead"
                  },
                  "title": "Response List Projects Api V1 Core Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/work-items": {
      "get": {
        "tags": [
          "core"
        ],
        "summary": "List Work Items",
        "operationId": "list_work_items_api_v1_core_work_items_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItemRead"
                  },
                  "title": "Response List Work Items Api V1 Core Work Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeQualityAnalyzeRequest": {
        "properties": {
          "repository_slug": {
            "type": "string",
            "title": "Repository Slug"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "default": "develop"
          },
          "scanners": {
            "items": {
              "type": "string",
              "enum": [
                "semgrep",
                "tree_sitter",
                "gitleaks",
                "trivy",
                "owasp_dependency_check"
              ]
            },
            "type": "array",
            "title": "Scanners"
          },
          "repository_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Path"
          },
          "execute_scanners": {
            "type": "boolean",
            "title": "Execute Scanners",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "repository_slug"
        ],
        "title": "CodeQualityAnalyzeRequest"
      },
      "CodeQualityAnalyzeResponse": {
        "properties": {
          "repository_slug": {
            "type": "string",
            "title": "Repository Slug"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "quality_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Quality Score"
          },
          "breakdown": {
            "$ref": "#/components/schemas/CodeQualityScoreBreakdown"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ScannerFinding"
            },
            "type": "array",
            "title": "Findings"
          },
          "evidence_policy": {
            "type": "string",
            "title": "Evidence Policy"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "repository_slug",
          "branch",
          "quality_score",
          "breakdown",
          "findings",
          "evidence_policy"
        ],
        "title": "CodeQualityAnalyzeResponse"
      },
...
      "CoverageVerificationRequest": {
        "properties": {
          "work_item_key": {
            "type": "string",
            "title": "Work Item Key"
          },
          "linked_commits": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Linked Commits",
            "default": 0
          },
          "linked_pull_requests": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Linked Pull Requests",
            "default": 0
          },
          "linked_test_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Linked Test Runs",
            "default": 0
          },
          "successful_pipeline_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Successful Pipeline Runs",
            "default": 0
          },
          "coverage_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Percent"
          }
        },
        "type": "object",
        "required": [
          "work_item_key"
        ],
        "title": "CoverageVerificationRequest"
      },
      "CoverageVerificationResponse": {
        "properties": {
          "work_item_key": {
            "type": "string",
            "title": "Work Item Key"
          },
          "coverage_confidence_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Coverage Confidence Score"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/CoverageGap"
            },
            "type": "array",
            "title": "Gaps"
          },
          "evidence_label": {
            "type": "string",
            "title": "Evidence Label",
            "default": "actual_or_placeholder_inputs"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "work_item_key",
          "coverage_confidence_score",
          "gaps"
        ],
        "title": "CoverageVerificationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          },
          "key": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Key"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "key"
        ],
        "title": "ProjectCreate"
      },
...
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "key"
        ],
        "title": "ProjectRead"
      },
...
      "RequirementGapInput": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "issue_key": {
            "type": "string",
            "title": "Issue Key"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "acceptance_criteria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Criteria"
          },
          "comments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "project_id",
          "issue_key",
          "summary"
        ],
        "title": "RequirementGapInput"
      },
      "RequirementGapOutput": {
        "properties": {
          "issue_key": {
            "type": "string",
            "title": "Issue Key"
          },
          "quality_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Quality Score"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/AiFindingOutput"
            },
            "type": "array",
            "title": "Findings"
          },
          "ai_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Run Id"
          },
          "finding_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Finding Ids"
          }
        },
        "type": "object",
        "required": [
          "issue_key",
          "quality_score",
          "findings"
        ],
        "title": "RequirementGapOutput"
      },
...
      "TenantCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Slug"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Calcutta"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "TenantCreate"
      },
      "TenantRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "timezone",
          "is_active"
        ],
        "title": "TenantRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkItemRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "issue_key": {
            "type": "string",
            "title": "Issue Key"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "source",
          "external_id",
          "issue_key",
          "issue_type",
          "status",
          "summary"
        ],
        "title": "WorkItemRead"
      }
    }
  }
}